requirements.txt
PyYAML==6.0.1
openpyxl==3.1.2
atlassian-python-api==03.41.11
click==8.1.7